monitoring/requirements.txt
#python3.11

numpy==2.2.2
pandas==2.2.3
polars==1.21.0
notebook==7.3.2
plotly==6.0.0
rerun-sdk==0.22.0
